Partie 4 fini (vérifier avec colin quand meme): le probleme était que les strength des animaux sont des int et dans la app.json nous les avons mis comme double ce qui ne faisais baisser les points de vies de personnes

+++ src/Environment/AntSoldier.cpp
@@ -27,7 +27,7 @@ sf::Sprite AntSoldier::getSprite() const
                        getDirection()/DEG_TO_RAD);
 }
 
-double AntSoldier::getStrength() const
+int AntSoldier::getStrength() const
 {
     return getAppConfig().ant_soldier_strength;
 }
