+++ src/Environment/AntSoldier.hpp
@@ -35,7 +35,7 @@ public:
      */
     sf::Sprite getSprite() const override;
 
-    double getStrength() const override;
+    int getStrength() const override;
 
     void drawOn(sf::RenderTarget& target) const override;
 };
